+++ src/main/kotlin/com/memorizy/server/controller/StudySetController.kt
@@ -5,6 +5,8 @@ import com.memorizy.server.service.StudySetService
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 
+// Controller, the task is to accept the request, send it to the correct service, receive a response and send it back.
+
 @RestController
 @RequestMapping("/api/sets")
 class StudySetController(
@@ -22,8 +24,8 @@ class StudySetController(
     }
 
     @DeleteMapping("/{id}")
-    fun deleteSet(@PathVariable id: Int): ResponseEntity<Void> {
+    fun deleteSet(@PathVariable id: Long): ResponseEntity<Void> {
         studySetService.deleteStudySet(id)
-        return ResponseEntity.noContent().build() // Возвращаем 204 No Content
+        return ResponseEntity.noContent().build()   // Возвращаем 204 No Content
     }
 }
