+++ src/main/kotlin/com/memorizy/server/dto/CardDto.kt
@@ -1,8 +1,10 @@
 package com.memorizy.server.dto
 
+// DTO (Data Transfer Object) - objects between client and server
+
 data class CardDto (
-    val id: Int? = null,
+    val id: Long? = null,
     val term: String,
     val definition: String,
-    val studySetId: Int
+    val studySetId: Long
 )
