+++ src/main/kotlin/com/memorizy/server/dto/StudySetDto.kt
@@ -1,7 +1,9 @@
 package com.memorizy.server.dto
 
+// DTO (Data Transfer Object) - objects between client and server
+
 data class StudySetDto(
-    val id: Int? = null, // ID может не быть при создании
+    val id: Long? = null, // ID может не быть при создании
     val name: String,
     val description: String?,
     val iconId: Int
