+++ src/main/kotlin/com/memorizy/server/model/Card.kt
@@ -2,12 +2,15 @@ package com.memorizy.server.model
 
 import jakarta.persistence.*
 
+// The entity for which Hibernate will create the tables itself
+
 @Entity
 @Table(name = "cards")
 data class Card(
+
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Int = 0,
+    @GeneratedValue(strategy = GenerationType.IDENTITY) // Assign database identity column
+    val id: Long = 0,
 
     @Column(nullable = false)
     val term: String,
@@ -15,6 +18,8 @@ data class Card(
     @Column(nullable = false)
     val definition: String,
 
+    // Много карточек может принадлежать одному набору
+    // FetchType.LAZY - если берем карточку из БД не грузим сразу информацию о наборе
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "study_set_id", nullable = false)
     val studySet: StudySet
